+++ src/main/java/com/umc/finly/domain/home/service/HomeMindServiceImpl.java
@@ -4,15 +4,12 @@ import com.umc.finly.domain.analysis.association.entity.ConvictionScoreResult;
 import com.umc.finly.domain.analysis.association.entity.FearIndexResult;
 import com.umc.finly.domain.home.dto.res.HomeMindDetailResDTO;
 import com.umc.finly.domain.home.dto.res.HomeMindResDTO;
-import com.umc.finly.domain.home.exception.code.HomeErrorCode;
+import com.umc.finly.domain.home.exception.HomeErrorCode;
 import com.umc.finly.domain.home.repository.HomeMindRepository;
 import com.umc.finly.domain.member.entity.Member;
 import com.umc.finly.domain.member.entity.Persona;
 import com.umc.finly.domain.member.repository.MemberRepository;
 import com.umc.finly.domain.member.repository.PersonaRepository;
-import com.umc.finly.domain.record.entity.RecordEntry;
-import com.umc.finly.domain.record.enums.EmotionCode;
-import com.umc.finly.domain.record.enums.TradeAction;
 import com.umc.finly.global.apiPayload.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
