+++ src/main/java/com/umc/finly/domain/market/exception/MarketException.java
@@ -1,6 +1,5 @@
 package com.umc.finly.domain.market.exception;
 
-import com.umc.finly.domain.market.exception.code.MarketErrorCode;
 import com.umc.finly.global.apiPayload.exception.CustomException;
 
 public class MarketException extends CustomException {
